experimenting
Currently the "model" does not have language-specific numerical
datatypes, this seems like a nice thing, as the code-emitter for a given
language can then map the numerical type to how it is best represented.
This should be extended for strings, pointers, and structs.
With this change, then the experiment here is to utlize jinja-filters to
materialize the type. Thus in the language-template one does:

* {{ entity | dtype_to_ctype }}

Following this other languages would have e.g. dtype_to_rtype (Rust),
dtype_to_ptype (Python), when writing this I realize that dtype_to_c,
dtype_to_rust, dtype_to_python would probably be must better names.

This attempts to settle on a couple of functions for the emission-work.
That is: emit_api and emit_pp. That is, the code generated will prodoce
the API and the pretty-print helpers. Once, FFI is added then these
tasks will most likely be expanded.

Are models, templates, and meta something that should be provided with
yace, or is that entirely up to the user? Hmm, it would make a lot of
sense that yace provides:

* An example model
* C templates to emit library, pretty-printers, and test-programs
* <lang> templates for FFI with the C library

All of these things such that the plumbing of yace itself can be tested,
independently of a specific model. This should also go as a means of
documenting the model and the possible entities within.

Signed-off-by: Simon A. F. Lund <os@safl.dk>