experimenting some more
Documentation: using sphinx-doc to document the project, using regular
prose describing the project along with a tutorial introduction and
"autodoc" to document the data-model.

Interface-Model: extended the plain YAML files and in-compiler-dict with
Python dataclasses. This is really convenient as the interface-model
needs documentation and verification. Thus, we the sphinx-doc / autodoc
it is very convenient to type up descriptions of the various
interface-model entities along with basic verification of members via
dataclass/typing.

Targets: a missing functionality is the definition of "targets"
e.g. what should be implemented to add a new code-emission target for
the yace compiler. One thing that needs to be thought into this is the
use of other other external tools e.g. generating documentation of the C
API, requires definition of e.g. Doxygen configuration file and then
invoking doxygen.

Models: with the addition of dataclasses several inconsistenties were
found and fixed. Additionally, the 'ltype' key is renamed 'cls' as it
serves the Python class encapsulation better.

General: 'ltype' renamed to 'cls' and 'topic' renamed to 'entity'.

Templates:

* add _str(),_pr(), and fpr() declr/def
* fix enter/exit struct+union
* use 'entity' for the primary/current entity for which code is emitted
  - Previously this was named 'topic'
* use 'member' when iterating over 'entity.members'

Signed-off-by: Simon A. F. Lund <os@safl.dk>