Re-organization and tool simplification.
setup
-----

Use 'package_data' instead of 'include_package_data'.

README / Documentation
----------------------

Major overhaul, simplified sections and added some text where missing.
It is shaping up nicely. The contents of README.rst is now equivalent to
the documentation landing page. An additional thing to do here, is to
ensure that the same content is provided as the package-readme on PyPi.

Templates
---------

These are now provided with the package instead of stray files. This
constributes to the simplification of the command-line interface. Down
the road when multiple targets are implemented, then templates are
problable something to package along with each yace-emitter-target
("yet"; thats a great name btw).

Yace-Interface-Model files
--------------------------

Previously a single module consisted of a collection of YAML files. To
make provide a simpler means of giving input them it is not expected
that a single yaml file contains the entire interface-model.
Additionally, a mandatory key "meta" is introduced. It replaces the need
for a separate "meta.yaml".

Yace-Interface-Model
--------------------

The Python modeling now has encapsulation of literals. A couple of
construction shorthands are added in addition to the default 'cls' for
each 'top-level' document. These are:

*dtype*: instead of providing the full:

  dtype: {cls: 'u32'}

then it now suffices to provide:

  dtype: u32

*lit*: instead of providing the full:

  lit: {cls: dec, val: d}

then it now suffices to provide:

  lit: 30

In both cases the InterfaceModel.entity_from_data() constructs the
corresponding datatypes.U32 and literals.LiteralDec entities.

Command-Line-Interface
----------------------

Richly simplified, now, only a single file is required: the YIM.

Misc
----

Added .idea to git-ignore.

TODO
----

* Fix the models, they need updating...
* Add a yim-format version number, it should be possible to produce
  transformation functions for one yim-format to another
* Add option ``--visualize`` to produce a neat visualization of the
  yim-tree

Signed-off-by: Simon A. F. Lund <os@safl.dk>