re-organize: move emitters to emitters and add function types
The "yets" (Yace Emitter Targets) need some organization, it is still
bundled in the yace package, eventually they should be able to live
externally, e.g. with packages:

  * yace-yet-<LANGUAGE> / yace-yet-<TARGET>

The targets could be the same language but variation on style /
conventions used in the generated code. Regardless, first step is to
get a bit clsoer at an interface for the yet-code. Now, it is provided
verbatim in `yace.emitters` with templates in `yace.templates`.

This commit also add the function types, except support for
function-pointers. Function-pointers is the last entity which is missing
representation in the IDL.

Signed-off-by: Simon A. F. Lund <os@safl.dk>