re-organize emitters / targets and cli
Emitters are modeled as the **things** which implement the
"transformation" of Entities to source-code-text, such as using Jinja
templates, a node-visitor, or some other technique.

**targets** are a combination of tasks: code-emitter, code-formatter,
and code-checker. Thus, a **target* has-a / uses one or more emitter
emitters to do what it does.

Before this re-organization them emitters where both the "technique" and
the specific target. This changes it into two separate **things**.

While doing this, some additional observations where made on
functionality which would make **yace** more useful, these where added
to the todo/roadmap.

Additionaly: the capi-target now invokes clang-format. The use of
clang-format is thus removed elsewhere in the project as it is now port
of ``yace.tools`` and used by ``yace.targets.capi``.

docs: show example with and without code-formating.

Signed-off-by: Simon A. F. Lund <os@safl.dk>