Refactor and add a code-path for linting
* Rename InterfaceModel to Model

* Move _traverse() / walk() code out of Model and into a re-usable
  ModelWalker().

Linting works on the loaded YIM, it works independently from what a
given target does. A target could utilize the linter when doing things
like linting the YIM after doing target-specific transformations.

Where should the code live which does the integrity-check of a given
Entity? Seemingly, the easiest is to add this logic with the Entity
itself, that is as a method named is_valid() on each data-classes in the
Model.

The linter task is thus reduced to walking the InterfaceModel, and
evaluating check() methods on the dataclasses. To help with this, then
yace.model.interface now has a class walking the InterfaceModel, for
things like the Linter can inherit.

model: replace typing.Union with tuple, since typing.Union is not
support for these purposes until Python 3.10.

Signed-off-by: Simon A. F. Lund <os@safl.dk>