mk,tests,cli,compiler: encapsulate stages in Compiler class
Ultimately an Encapsulating Compiler class should exist. Adding it now
to provide an entry-point into the code for testing, specifically, to do
negative-tests, such as hitting exception-paths etc.

The above is utilized by the pytest in 'tests' and the Makefile-helper
for coverage reporting is extended to make use of this as well. Thus
covering both the cli-invocation as well as the programmatic invocation
of the Yace compiler.

Signed-off-by: Simon A. F. Lund <os@safl.dk>