Replace --stage with --lint
The '--stages' argument to the yace-cli, provided a means for the user
to control with compiler-stages the given model should be processed
under. However, since only a few set of stages make sense:

1) parse, lint
2) parse, lint, emit
3) parse, lint, emit, format
4) parse, lint, emit, format, check

And of the above, only 1) and 4) would be considered common use-cases,
the other as means to "test" the compiler, not actual usage. Thus, 4) is
the default cli-behavior and by supplying '--lint' default behavior is
changed to only do the stages in 1).

For test-instrumentation, then then 'stages' argument is moved from the
Compiler-constructor to the Compiler.process(path, stages).

Signed-off-by: Simon A. F. Lund <os@safl.dk>