Reduce model-coupling and add general target-initialization
The model was tightly coupled to the Compiler, Emitter, Target, and
classes inheriting from these via init. There where no reason for these
to have the model as constructor argument, thus removing it, and
adding it as method argument where applicable.

The ABC-class Target, now has a more useful initializer function, this
provides for potentially slightly more alignment on different
target-implementations. E.g. a mixin doing something can rely on the
features of the Target rather than on features of a specific
Target-implementation.

Signed-off-by: Simon A. F. Lund <os@safl.dk>