emitter/jinja: restructure for separation and re-use
The content of the Jinja-templates mixes language syntax with the
Jinja-logic with the escape-chars: {{, {%, {%-. As such, this is a bit
hard on the eyes, additionally the reader has to keep two language
syntaxes in their mind at the same time. This is taxing.

The changes here is an attempt at aleviating this a bit, by separarting
the emitter-logic out into more files.

Signed-off-by: Simon A. F. Lund <os@safl.dk>