model/entity,models/nvme: add validation dispatch for attr
The `Entity` class can validate the simple types of attributes, however,
specific charactaristics should be validated by a function in the class
itself, such as ``is_valid_sym`` for the attribute named ``sym`` in the
mixin-class adding the attribute.

This change does so.

Additionally, then the nvme-model is adjusted to pass the
validation-checks.

Signed-off-by: Simon A. F. Lund <os@safl.dk>