docs,setup,parser,cli/yace: add a C parser and cli-tool to use it
Producing yidl-files manually is fine when starting from scratch.
However, for an existing project, then a C header file usually exists.
By parsing it, most of the yidl-boiler-plate can be produced.

The addition of a C parser is added for this purpose, to support the
addition of cli-tool to parse C to emit yidl.

* setup: add 'libclang' to requirements, as the parser uses it
* docs: add a section in the codebase for describing the parser
* parser: add the CParser itself
* cli/yace: add use of the CParser to produce YIDL

This is all done in preparation for the tool which will use this parser.

short-hand abbreviation
-----------------------

When construcing typespec data, the full expanded form is created. When
dumping the expanded form to file then it is not very readable. This is
why the yidl has "shorthands". To make good use of these, then this
change adds a conversion of datatypes to shorthand abbreviation where
possible.

Signed-off-by: Simon A. F. Lund <os@safl.dk>