targets/capi: use self.NAME instead of CAPI.NAME
It would make sense that it should be simple to create targets which are
derivates of the "builtin" 'capi' target. E.g. produce a 'capi' + some
additional helper-functions / utilities etc.

By replace CAPI.NAME with self.NAME, then it should be possible to do
just that. However, the addition of new jinja2-templates would need an
additional emitter instance etc.

Signed-off-by: Simon A. F. Lund <os@safl.dk>