emitters,targets: replace PackageLoader with FileSystemLoader
yace.emitter.Emitter uses the PackageLoader, needless to say, this
requires that the template is inside a package. This is a downside when
using adhoc-targets, these cannot use the Emitter, since they are not
packages.
Thus, to enable this usecase then the FileSystemLoader is used instead.
The two other constructor-arguments were mainly unused anyway, thus this
seems to be an overall improvement of the Emitter.

Signed-off-by: Simon A. F. Lund <os@safl.dk>