setup: move package info. from setup.py to setup.cfg
The Python community is aiming to deprecate the legacy approach to
project metadata and packaging. An intermediate replacement is setup.cfg
and the end-goal is pyproject.toml. The pyproject.toml is currently not
fully supported by e.g. setuptools, e.g. some functionality are still in
beta.
Thus, for yace, the following makes sense:

* provide setup.py as legacy build-usage
* provide setup.cfg for all project meta-data and packaging information
* provide pyproject.toml for build-system and tool options

Signed-off-by: Simon A. F. Lund <os@safl.dk>