refactor(emitters,capi): use c_spelling() intead of templ.
The emitter for a typespec was in the CAPI handled by a typespec
template. This is replaced with a call to 'entity.typ.c_spelling()'.

The CAPI emitter templates had become complex. While this isn't
inherently a problem, the author struggled to read the code. The
purpose of using templates was, and still is, to create easy-to-read and
easy-to-write emitters, thereby reducing the complexity of the emitter
templates.

What is lost is the ability to emit nested definitions in structs/
unions. However, since this feature is not supported by design, there
is no reason to retain code that emits such definitions, aside from some
form of pride.

Signed-off-by: Simon A. F. Lund <os@safl.dk>