feat(ir): add module attribute to entities and module model walker
If we expect the input file for yace to be a YAML file including all
entities of a full library, we need the notion of a "module" in the
internal representation such that we can produce meaningful files for
the given targets.

This commit adds the optional module attribute to all entities, and a
ModelWalker, which extracts the module from entity names and adds it to
the entity. This operates under the assumption that the given input
follows the convention of naming all entities in the format:
<prefix>_<module>_<entity_name>

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>